feat: added app routing
