+++ src/app/app.component.html
@@ -1,18 +1,3 @@
 <app-navbar></app-navbar>
-<ng-container *ngIf="ytVideos$ | async as ytVideo">
-  <div class="relative bg-gray-200 overflow-hidden">
-    <div class="max-w-7xl mx-auto py-8 px-4 sm:px-6 lg:px-8">
-      <app-video-banner
-        [ytVideo]="ytVideo.lastVideo || null"
-      ></app-video-banner>
-    </div>
-  </div>
-  <div class="relative bg-white overflow-hidden">
-    <main class="mx-auto max-w-7xl pb-8 pt-6 px-4 sm:px-6 lg:px-8">
-      <app-video-listing
-        [ytVideos]="ytVideo.videoList || null"
-      ></app-video-listing>
-    </main>
-  </div>
-</ng-container>
+<router-outlet></router-outlet>
 <app-footer></app-footer>
